TaskRemoveDirectory

+++ Oetools.Builder/Project/Task/OeTaskFileDelete.cs
@@ -29,21 +29,25 @@ using Oetools.Utilities.Lib.Extension;
 namespace Oetools.Builder.Project.Task {
     
     /// <summary>
-    /// This tasks simply delete file path (in a folder)
+    /// This tasks simply delete files path (in a folder)
     /// </summary>
     [Serializable]
     [XmlRoot("Delete")]
-    public class OeTaskFileDelete : OeTaskFile {   
+    public class OeTaskFileDelete : OeTaskFile {
         
         protected override void ExecuteForFilesInternal(FileList<OeFile> files) {
 
             foreach (var file in files) {
-                try {
-                    CancelSource?.Token.ThrowIfCancellationRequested();
-                    Log?.Trace?.Write($"Deleting file {file.SourcePathForTaskExecution.PrettyQuote()}");
-                    File.Delete(file.SourcePathForTaskExecution);
-                } catch (Exception e) {
-                    throw new TaskExecutionException(this, $"Could not delete {file.SourcePathForTaskExecution.PrettyQuote()}", e);
+                if (File.Exists(file.SourcePathForTaskExecution)) {
+                    try {
+                        CancelSource?.Token.ThrowIfCancellationRequested();
+                        Log?.Trace?.Write($"Deleting file {file.SourcePathForTaskExecution.PrettyQuote()}");
+                        File.Delete(file.SourcePathForTaskExecution);
+                    } catch (Exception e) {
+                        throw new TaskExecutionException(this, $"Could not delete file {file.SourcePathForTaskExecution.PrettyQuote()}", e);
+                    }
+                } else {
+                    Log?.Trace?.Write($"Deleting file not existing {file.SourcePathForTaskExecution.PrettyQuote()}");
                 }
             }
